Moved dev tasks files to their final location.
