--- dev/tasks/githooks.js
@@ -1,13 +1,0 @@
-'use strict';
-
-module.exports = function( grunt ) {
-	grunt.config.merge( {
-		githooks: {
-			all: {
-				'pre-commit': 'default'
-			}
-		}
-	} );
-
-	grunt.loadNpmTasks( 'grunt-githooks' );
-};
